% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_date.R
\name{conv_date}
\alias{conv_date}
\title{Convert a string or a POSIXt in a Date}
\usage{
conv_date(x, format = "\%d.\%m.\%Y")
}
\arguments{
\item{x}{a date in string or POSIXt format}

\item{format}{date format (for strings only)}
}
\value{
a date in the Date format
}
\description{
Convert a string or a POSIXt in a Date
}
\examples{

}
