% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_data.R
\name{daily_data}
\alias{daily_data}
\title{Generate daily data}
\usage{
daily_data(data, format = "wide")
}
\arguments{
\item{data}{data imported with import_data()}

\item{format}{either wide (default) or long}
}
\value{
a dataframe
variables in wide format:
\itemize{
  \item date     = date
  \item cumhos   = cumulative number of hospitalized patients
  \item hos      = current number of hospitalized patients
  \item icu      = current number of patients treated in intensive care
  \item rest     = current number of patients hospitalized after
                   intensive care
  \item death    = cumulative number of deaths
  \item recovery = cumulative number of recoveries
}
variables in long format:
\itemize{
  \item date     = date
  \item state    = hos, icu, rest, death or recovery
  \item n        = count
}
}
\description{
Generate daily data from REDCap data in wide or long format
}
\examples{

}
