% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import REDCap data}
\usage{
import_data(path, ..., censoring_date = NULL)
}
\arguments{
\item{path}{path to data file}

\item{...}{other parameter passed to readxl::read_xlsx}

\item{censoring_date}{date used for censoring (optional, format: \%Y-\%m-\%d)}
}
\value{
a dataframe containing the following variables
\itemize{
  \item id      = anonymized identifier
  \item sex     = patient's gender (factor M/F)
  \item age     = patient's age (years)
  \item hos_in  = date of hospitalization
  \item icu_in  = date of admission in ICU
  \item icu_out = intensive care discharge date
  \item hos_out = discharge date from hospital
  \item death   = death indicator (logical true/false)
}
The censoring date is returned as an attributes (censoring_date)
}
\description{
Import a formatted Excel REDCap datafile containing the following variables:
\itemize{
  \item no_unique            = anonymized identifier
  \item sex                  = patient's gender (M/F)
  \item age                  = patient's age (years)
  \item arrivee_hopital      = date of hospitalization
  \item debut_soins_intensif = date of admission in ICU
  \item fin_soins_intensifs  = intensive care discharge date
  \item sortie_hopital       = discharge date from hospital
  \item deces                = death indicator (0/1)
}
The four date variable must be in date format in Excel. A censoring date
could be defined. If such a date is not given, it is guessed based on the
data.
}
\examples{

}
