% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_occupancy.R
\name{plot_occupancy}
\alias{plot_occupancy}
\title{Plot state occupancy}
\usage{
plot_occupancy(
  pred,
  state = "all",
  pi = FALSE,
  pi_length = 0.9,
  ttl = "",
  sttl = ""
)
}
\arguments{
\item{pred}{an object returned by pred_state_occupancy}

\item{state}{a vector containing one or more of the following values: 'hos',
'icu', 'rest', 'death', 'recovery'; use 'all' to select all the states}

\item{pi}{a logical value (TRUE/FALSE) indicating if prediction intervall
must be drawn}

\item{pi_length}{length of predicion intervall (ignored if pi = FALSE)}

\item{ttl}{plot title}

\item{sttl}{plot subtitle}
}
\value{
a plot of state occupancy
}
\description{
Plot state occupancy for 1.25 multi-state model
}
\examples{

}
