% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_state_occupancy.R
\name{pred_state_occupancy}
\alias{pred_state_occupancy}
\title{Predict state occupancy}
\usage{
pred_state_occupancy(
  fit,
  trans,
  nday,
  nsim,
  pars,
  data,
  start_date = "2020-02-25",
  seed = 666,
  ncpu = 1
)
}
\arguments{
\item{fit}{multi-state model (or list of models, one per possible
transition) estimated with flexsurv::flexsurvreg}

\item{trans}{matrix describing the possible transitions}

\item{nday}{number of days to predict}

\item{nsim}{number of simulations}

\item{pars}{table of parameter}

\item{data}{data imported with import_data()}

\item{start_date}{date to start simulations}

\item{seed}{seed of the random number generator for reproducible analyses}

\item{ncpu}{number of cpus to use for parallel computations}
}
\value{
a list of simultations (each element of the list correspond to a
different state); the data are also returned as an attribute (data)
}
\description{
Predict state occupancy for 1.25 multi-state model
}
\examples{

}
