% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepare data for mutli-state analysis}
\usage{
prep_data(data, format = "mstate")
}
\arguments{
\item{data}{data imported with import_data()}

\item{format}{either survival or mstate (default)}
}
\value{
a dataframe built with survival::tmerge or mstate::mprep; if format
is 'mstate', then the matrix of possible transitions (tmat) is returned as
an attribute
}
\description{
Prepare data for a multi-state analysis either with survival::survfit or
mstate and/or flexsurvreg
}
\examples{

}
