% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regp.R
\name{regp}
\alias{regp}
\title{Random sample for exponential growth parameter}
\usage{
regp(n, megp, vegp)
}
\arguments{
\item{n}{number of draws}

\item{megp}{median of growth parameter (typically 1.25)}

\item{vegp}{variability}
}
\value{
a vector sampled from the exponential growth parameter distribution
}
\description{
Random sample for exponential growth parameter (constrained >1)
}
\examples{

}
