% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_fmsm_traj.R
\name{sim_fmsm_traj}
\alias{sim_fmsm_traj}
\title{Generate trajectories for given covariates}
\usage{
sim_fmsm_traj(..., tvec)
}
\arguments{
\item{...}{arguments passed to sim.fmsm (exept t which is calculated from
tvec)}

\item{tvec}{a vector of times for which the states have to be determined}
}
\value{
a matrix of states, each row coresponds to an individual and each
column to a time
}
\description{
Generate trajectories from a semi-Markov model built with
flexsurv::flexsurvreg. The trajectories are simulate with
flexsurv::sim.fmsm. The latter return the times of the transitions.  This
function determines the states M for each individual for each time given in
an input vector.
}
\examples{

}
