% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_state_occupancy.R
\name{sim_state_occupancy}
\alias{sim_state_occupancy}
\title{Simulate state occupancy}
\usage{
sim_state_occupancy(fit, trans, newdata, nday)
}
\arguments{
\item{fit}{multi-state model (or list of models, one per possible
transition) estimated with flexsurv::flexsurvreg}

\item{trans}{matrix describing the possible transitions}

\item{newdata}{dataframe containing the covariates of the new patients (one
row per new patient)}

\item{nday}{number of days to simulate}
}
\value{
a matrix containing the number of patients per state (rows) and per
day (cols)
}
\description{
Simulate state occupancy of new patients for a given number of days.
}
\examples{

}
