% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_to_traj.R
\name{sim_to_traj}
\alias{sim_to_traj}
\title{Generate trajectories from flexsurv::sim.fmsm}
\usage{
sim_to_traj(sim, tvec)
}
\arguments{
\item{sim}{output of flexsurv::sim.fmsm}

\item{tvec}{a vector of times for which the states have to be determined}
}
\value{
a matrix of states, each row coresponds to an individual and each
column to a time
}
\description{
Generate trajectories from the simulations generated by flexsurv::sim.fmsm
for a semi-Markov model built with flexsurv::flexsurvreg. The function
flexsurv::sim.fmsm returns the times of the transitions.  This
function determines each time given in the vector tvec.
}
\examples{

}
